refactor: use-account instead of useAccount
